feat: implement pagination for ListConnections and ListAgents endpoints
- Updated ListConnections and ListAgents methods to accept pagination options (limit and continue token).
- Modified the controller, service, and use case layers to handle the new pagination logic.
- Enhanced error handling for invalid limit parameters in ListConnections.
- Updated API documentation to reflect new query parameters for pagination.
- Added KeyValues method to MultiMap to facilitate key-value retrieval.

+++ api/v1/v1.go
@@ -7,7 +7,8 @@ const (
 	APIVersion = "v1"
 )
 
+// ListMeta is a struct that contains metadata for list responses.
 type ListMeta struct {
 	Continue           string `json:"continue"`
-	RemainingItemCount int    `json:"remainingItemCount"`
+	RemainingItemCount int64  `json:"remainingItemCount"`
 } // @name ListMeta
